--- x/ccv/provider/keeper/legacy_proposal.go
@@ -1,35 +1,0 @@
-package keeper
-
-import (
-	sdk "github.com/cosmos/cosmos-sdk/types"
-
-	"github.com/cosmos/interchain-security/v5/x/ccv/provider/types"
-)
-
-func (k Keeper) HandleLegacyConsumerRewardDenomProposal(ctx sdk.Context, p *types.ChangeRewardDenomsProposal) error {
-	for _, denomToAdd := range p.DenomsToAdd {
-		// Log error and move on if one of the denoms is already registered
-		if k.ConsumerRewardDenomExists(ctx, denomToAdd) {
-			ctx.Logger().Error("denom %s already registered", denomToAdd)
-			continue
-		}
-		k.SetConsumerRewardDenom(ctx, denomToAdd)
-		ctx.EventManager().EmitEvent(sdk.NewEvent(
-			types.EventTypeAddConsumerRewardDenom,
-			sdk.NewAttribute(types.AttributeConsumerRewardDenom, denomToAdd),
-		))
-	}
-	for _, denomToRemove := range p.DenomsToRemove {
-		// Log error and move on if one of the denoms is not registered
-		if !k.ConsumerRewardDenomExists(ctx, denomToRemove) {
-			ctx.Logger().Error("denom %s not registered", denomToRemove)
-			continue
-		}
-		k.DeleteConsumerRewardDenom(ctx, denomToRemove)
-		ctx.EventManager().EmitEvent(sdk.NewEvent(
-			types.EventTypeRemoveConsumerRewardDenom,
-			sdk.NewAttribute(types.AttributeConsumerRewardDenom, denomToRemove),
-		))
-	}
-	return nil
-}
