+++ x/ccv/provider/keeper/permissionless_test.go
@@ -4,14 +4,9 @@ import (
 	"testing"
 	"time"
 
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	govv1 "github.com/cosmos/cosmos-sdk/x/gov/types/v1"
-	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 	"github.com/cosmos/ibc-go/v8/modules/core/02-client/types"
 	testkeeper "github.com/cosmos/interchain-security/v5/testutil/keeper"
-	"github.com/cosmos/interchain-security/v5/x/ccv/provider/keeper"
 	providertypes "github.com/cosmos/interchain-security/v5/x/ccv/provider/types"
-	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/require"
 )
 
@@ -244,149 +239,6 @@ func TestConsumerPhase(t *testing.T) {
 	require.Equal(t, providertypes.ConsumerPhase_CONSUMER_PHASE_LAUNCHED, phase)
 }
 
-// TestConsumerStopTime tests the getter, setter, and deletion of the consumer id to stop times methods
-func TestConsumerStopTime(t *testing.T) {
-	providerKeeper, ctx, ctrl, _ := testkeeper.GetProviderKeeperAndCtx(t, testkeeper.NewInMemKeeperParams(t))
-	defer ctrl.Finish()
-
-	_, err := providerKeeper.GetConsumerStopTime(ctx, "consumerId")
-	require.Error(t, err)
-
-	expectedStopTime := time.Unix(1234, 56789)
-	providerKeeper.SetConsumerStopTime(ctx, "consumerId", expectedStopTime)
-	actualStopTime, err := providerKeeper.GetConsumerStopTime(ctx, "consumerId")
-	require.NoError(t, err)
-	require.Equal(t, actualStopTime, expectedStopTime)
-
-	providerKeeper.DeleteConsumerStopTime(ctx, "consumerId")
-	_, err = providerKeeper.GetConsumerStopTime(ctx, "consumerId")
-	require.Error(t, err)
-}
-
-// TestConsumersToBeLaunched tests `AppendConsumerToBeLaunched`, `GetConsumersToBeLaunched`, and `RemoveConsumerToBeLaunched`
-func TestConsumersToBeLaunched(t *testing.T) {
-	providerKeeper, ctx, ctrl, _ := testkeeper.GetProviderKeeperAndCtx(t, testkeeper.NewInMemKeeperParams(t))
-	defer ctrl.Finish()
-
-	spawnTime := time.Now()
-	providerKeeper.AppendConsumerToBeLaunched(ctx, "consumerId1", spawnTime)
-	consumers, err := providerKeeper.GetConsumersToBeLaunched(ctx, spawnTime)
-	require.NoError(t, err)
-	require.Equal(t, []string{"consumerId1"}, consumers.Ids)
-
-	providerKeeper.AppendConsumerToBeLaunched(ctx, "consumerId2", spawnTime)
-	consumers, err = providerKeeper.GetConsumersToBeLaunched(ctx, spawnTime)
-	require.NoError(t, err)
-	require.Equal(t, []string{"consumerId1", "consumerId2"}, consumers.Ids)
-
-	providerKeeper.AppendConsumerToBeLaunched(ctx, "consumerId3", spawnTime)
-	consumers, err = providerKeeper.GetConsumersToBeLaunched(ctx, spawnTime)
-	require.NoError(t, err)
-	require.Equal(t, []string{"consumerId1", "consumerId2", "consumerId3"}, consumers.Ids)
-
-	err = providerKeeper.RemoveConsumerToBeLaunched(ctx, "consumerId2", spawnTime)
-	require.NoError(t, err)
-	consumers, err = providerKeeper.GetConsumersToBeLaunched(ctx, spawnTime)
-	require.NoError(t, err)
-	require.Equal(t, []string{"consumerId1", "consumerId3"}, consumers.Ids)
-
-	// also add consumer ids under a different spawn time and verify everything under the original spawn time is still there
-	spawnTimePlusOneHour := spawnTime.Add(time.Hour)
-	providerKeeper.AppendConsumerToBeLaunched(ctx, "consumerId4", spawnTimePlusOneHour)
-	consumers, err = providerKeeper.GetConsumersToBeLaunched(ctx, spawnTimePlusOneHour)
-	require.NoError(t, err)
-	require.Equal(t, []string{"consumerId4"}, consumers.Ids)
-
-	consumers, err = providerKeeper.GetConsumersToBeLaunched(ctx, spawnTime)
-	require.NoError(t, err)
-	require.Equal(t, []string{"consumerId1", "consumerId3"}, consumers.Ids)
-
-	// start removing all consumers from `spawnTime`
-	err = providerKeeper.RemoveConsumerToBeLaunched(ctx, "consumerId3", spawnTime)
-	require.NoError(t, err)
-	err = providerKeeper.RemoveConsumerToBeLaunched(ctx, "consumerId1", spawnTime)
-	require.NoError(t, err)
-	consumers, err = providerKeeper.GetConsumersToBeLaunched(ctx, spawnTime)
-	require.NoError(t, err)
-	require.Empty(t, consumers.Ids)
-
-	// remove from `spawnTimePlusOneHour`
-	err = providerKeeper.RemoveConsumerToBeLaunched(ctx, "consumerId4", spawnTimePlusOneHour)
-	require.NoError(t, err)
-	consumers, err = providerKeeper.GetConsumersToBeLaunched(ctx, spawnTimePlusOneHour)
-	require.NoError(t, err)
-	require.Empty(t, consumers.Ids)
-
-	// add another consumer for `spawnTime`
-	err = providerKeeper.AppendConsumerToBeLaunched(ctx, "consumerId5", spawnTime)
-	require.NoError(t, err)
-	consumers, err = providerKeeper.GetConsumersToBeLaunched(ctx, spawnTime)
-	require.NoError(t, err)
-	require.Equal(t, []string{"consumerId5"}, consumers.Ids)
-}
-
-// TestConsumersToBeStopped tests `AppendConsumerToBeLaunched`, `GetConsumersToBeLaunched`, and `RemoveConsumerToBeLaunched`
-func TestConsumersToBeStopped(t *testing.T) {
-	providerKeeper, ctx, ctrl, _ := testkeeper.GetProviderKeeperAndCtx(t, testkeeper.NewInMemKeeperParams(t))
-	defer ctrl.Finish()
-
-	stopTime := time.Now()
-	providerKeeper.AppendConsumerToBeStopped(ctx, "consumerId1", stopTime)
-	consumers, err := providerKeeper.GetConsumersToBeStopped(ctx, stopTime)
-	require.NoError(t, err)
-	require.Equal(t, []string{"consumerId1"}, consumers.Ids)
-
-	providerKeeper.AppendConsumerToBeStopped(ctx, "consumerId2", stopTime)
-	consumers, err = providerKeeper.GetConsumersToBeStopped(ctx, stopTime)
-	require.NoError(t, err)
-	require.Equal(t, []string{"consumerId1", "consumerId2"}, consumers.Ids)
-
-	providerKeeper.AppendConsumerToBeStopped(ctx, "consumerId3", stopTime)
-	consumers, err = providerKeeper.GetConsumersToBeStopped(ctx, stopTime)
-	require.NoError(t, err)
-	require.Equal(t, []string{"consumerId1", "consumerId2", "consumerId3"}, consumers.Ids)
-
-	err = providerKeeper.RemoveConsumerToBeStopped(ctx, "consumerId2", stopTime)
-	require.NoError(t, err)
-	consumers, err = providerKeeper.GetConsumersToBeStopped(ctx, stopTime)
-	require.NoError(t, err)
-	require.Equal(t, []string{"consumerId1", "consumerId3"}, consumers.Ids)
-
-	// also add consumer ids under a different stop time and verify everything under the original stop time is still there
-	stopTimePlusOneHour := stopTime.Add(time.Hour)
-	providerKeeper.AppendConsumerToBeStopped(ctx, "consumerId4", stopTimePlusOneHour)
-	consumers, err = providerKeeper.GetConsumersToBeStopped(ctx, stopTimePlusOneHour)
-	require.NoError(t, err)
-	require.Equal(t, []string{"consumerId4"}, consumers.Ids)
-
-	consumers, err = providerKeeper.GetConsumersToBeStopped(ctx, stopTime)
-	require.NoError(t, err)
-	require.Equal(t, []string{"consumerId1", "consumerId3"}, consumers.Ids)
-
-	// start removing all consumers from `stopTime`
-	err = providerKeeper.RemoveConsumerToBeStopped(ctx, "consumerId3", stopTime)
-	require.NoError(t, err)
-	err = providerKeeper.RemoveConsumerToBeStopped(ctx, "consumerId1", stopTime)
-	require.NoError(t, err)
-	consumers, err = providerKeeper.GetConsumersToBeStopped(ctx, stopTime)
-	require.NoError(t, err)
-	require.Empty(t, consumers.Ids)
-
-	// remove from `stopTimePlusOneHour`
-	err = providerKeeper.RemoveConsumerToBeStopped(ctx, "consumerId4", stopTimePlusOneHour)
-	require.NoError(t, err)
-	consumers, err = providerKeeper.GetConsumersToBeStopped(ctx, stopTimePlusOneHour)
-	require.NoError(t, err)
-	require.Empty(t, consumers.Ids)
-
-	// add another consumer for `stopTime`
-	err = providerKeeper.AppendConsumerToBeStopped(ctx, "consumerId5", stopTime)
-	require.NoError(t, err)
-	consumers, err = providerKeeper.GetConsumersToBeStopped(ctx, stopTime)
-	require.NoError(t, err)
-	require.Equal(t, []string{"consumerId5"}, consumers.Ids)
-}
-
 // TestOptedInConsumerIds tests the `GetOptedInConsumerIds`, `AppendOptedInConsumerId`, and `RemoveOptedInConsumerId` methods
 func TestGetOptedInConsumerIds(t *testing.T) {
 	providerKeeper, ctx, ctrl, _ := testkeeper.GetProviderKeeperAndCtx(t, testkeeper.NewInMemKeeperParams(t))
@@ -441,192 +293,6 @@ func TestGetOptedInConsumerIds(t *testing.T) {
 	require.Empty(t, consumers)
 }
 
-// TestConsumerChainId tests the getter, setter, and deletion of the client id to consumer id methods
-func TestClientIdToConsumerId(t *testing.T) {
-	providerKeeper, ctx, ctrl, _ := testkeeper.GetProviderKeeperAndCtx(t, testkeeper.NewInMemKeeperParams(t))
-	defer ctrl.Finish()
-
-	_, found := providerKeeper.GetClientIdToConsumerId(ctx, "clientId")
-	require.False(t, found)
-
-	providerKeeper.SetClientIdToConsumerId(ctx, "clientId", "consumerId")
-	consumerId, found := providerKeeper.GetClientIdToConsumerId(ctx, "clientId")
-	require.True(t, found)
-	require.Equal(t, "consumerId", consumerId)
-
-	// assert that overwriting the current consumer id record works
-	providerKeeper.SetClientIdToConsumerId(ctx, "clientId", "consumerId2")
-	consumerId, found = providerKeeper.GetClientIdToConsumerId(ctx, "clientId")
-	require.True(t, found)
-	require.Equal(t, "consumerId2", consumerId)
-
-	providerKeeper.DeleteClientIdToConsumerId(ctx, "clientId")
-	consumerId, found = providerKeeper.GetClientIdToConsumerId(ctx, "clientId")
-	require.False(t, found)
-	require.Equal(t, "", consumerId)
-}
-
-// TestGetInitializedConsumersReadyToLaunch tests that the ready to-be-launched consumer chains are returned
-func TestGetInitializedConsumersReadyToLaunch(t *testing.T) {
-	providerKeeper, ctx, ctrl, _ := testkeeper.GetProviderKeeperAndCtx(t, testkeeper.NewInMemKeeperParams(t))
-	defer ctrl.Finish()
-
-	// no chains to-be-launched exist
-	require.Empty(t, providerKeeper.GetInitializedConsumersReadyToLaunch(ctx, 5))
-
-	providerKeeper.AppendConsumerToBeLaunched(ctx, "consumerId1", time.Unix(10, 0))
-	providerKeeper.AppendConsumerToBeLaunched(ctx, "consumerId2", time.Unix(20, 0))
-	providerKeeper.AppendConsumerToBeLaunched(ctx, "consumerId3", time.Unix(30, 0))
-
-	// time has not yet reached the spawn time of "consumerId1"
-	ctx = ctx.WithBlockTime(time.Unix(9, 999999999))
-	require.Empty(t, providerKeeper.GetInitializedConsumersReadyToLaunch(ctx, 3))
-
-	// time has reached the spawn time of "consumerId1"
-	ctx = ctx.WithBlockTime(time.Unix(10, 0))
-	require.Equal(t, []string{"consumerId1"}, providerKeeper.GetInitializedConsumersReadyToLaunch(ctx, 3))
-
-	// time has reached the spawn time of "consumerId1" and "consumerId2"
-	ctx = ctx.WithBlockTime(time.Unix(20, 0))
-	require.Equal(t, []string{"consumerId1", "consumerId2"}, providerKeeper.GetInitializedConsumersReadyToLaunch(ctx, 3))
-
-	// time has reached the spawn time of all chains
-	ctx = ctx.WithBlockTime(time.Unix(30, 0))
-	require.Equal(t, []string{"consumerId1", "consumerId2", "consumerId3"}, providerKeeper.GetInitializedConsumersReadyToLaunch(ctx, 3))
-
-	// limit the number of returned consumer chains
-	require.Equal(t, []string{}, providerKeeper.GetInitializedConsumersReadyToLaunch(ctx, 0))
-	require.Equal(t, []string{"consumerId1"}, providerKeeper.GetInitializedConsumersReadyToLaunch(ctx, 1))
-	require.Equal(t, []string{"consumerId1", "consumerId2"}, providerKeeper.GetInitializedConsumersReadyToLaunch(ctx, 2))
-}
-
-func TestGetLaunchedConsumersReadyToStop(t *testing.T) {
-	providerKeeper, ctx, ctrl, _ := testkeeper.GetProviderKeeperAndCtx(t, testkeeper.NewInMemKeeperParams(t))
-	defer ctrl.Finish()
-
-	// no chains to-be-stopped exist
-	require.Empty(t, providerKeeper.GetLaunchedConsumersReadyToStop(ctx, 3))
-
-	providerKeeper.AppendConsumerToBeStopped(ctx, "consumerId1", time.Unix(10, 0))
-	providerKeeper.AppendConsumerToBeStopped(ctx, "consumerId2", time.Unix(20, 0))
-	providerKeeper.AppendConsumerToBeStopped(ctx, "consumerId3", time.Unix(30, 0))
-
-	// time has not yet reached the stop time of "consumerId1"
-	ctx = ctx.WithBlockTime(time.Unix(9, 999999999))
-	require.Empty(t, providerKeeper.GetLaunchedConsumersReadyToStop(ctx, 3))
-
-	// time has reached the stop time of "consumerId1"
-	ctx = ctx.WithBlockTime(time.Unix(10, 0))
-	require.Equal(t, []string{"consumerId1"}, providerKeeper.GetLaunchedConsumersReadyToStop(ctx, 3))
-
-	// time has reached the stop time of "consumerId1" and "consumerId2"
-	ctx = ctx.WithBlockTime(time.Unix(20, 0))
-	require.Equal(t, []string{"consumerId1", "consumerId2"}, providerKeeper.GetLaunchedConsumersReadyToStop(ctx, 3))
-
-	// time has reached the stop time of all chains
-	ctx = ctx.WithBlockTime(time.Unix(30, 0))
-	require.Equal(t, []string{"consumerId1", "consumerId2", "consumerId3"}, providerKeeper.GetLaunchedConsumersReadyToStop(ctx, 3))
-}
-
-func TestUpdateAllowlist(t *testing.T) {
-	providerKeeper, ctx, ctrl, _ := testkeeper.GetProviderKeeperAndCtx(t, testkeeper.NewInMemKeeperParams(t))
-	defer ctrl.Finish()
-
-	consumerId := "0"
-
-	providerConsAddr1 := "cosmosvalcons1qmq08eruchr5sf5s3rwz7djpr5a25f7xw4mceq"
-	consAddr1, _ := sdk.ConsAddressFromBech32(providerConsAddr1)
-	providerConsAddr2 := "cosmosvalcons1nx7n5uh0ztxsynn4sje6eyq2ud6rc6klc96w39"
-	consAddr2, _ := sdk.ConsAddressFromBech32(providerConsAddr2)
-
-	providerKeeper.UpdateAllowlist(ctx, consumerId, []string{providerConsAddr1, providerConsAddr2})
-
-	expectedAllowlist := []providertypes.ProviderConsAddress{
-		providertypes.NewProviderConsAddress(consAddr1),
-		providertypes.NewProviderConsAddress(consAddr2)}
-	require.Equal(t, expectedAllowlist, providerKeeper.GetAllowList(ctx, consumerId))
-}
-
-func TestUpdateDenylist(t *testing.T) {
-	providerKeeper, ctx, ctrl, _ := testkeeper.GetProviderKeeperAndCtx(t, testkeeper.NewInMemKeeperParams(t))
-	defer ctrl.Finish()
-
-	consumerId := "0"
-
-	providerConsAddr1 := "cosmosvalcons1qmq08eruchr5sf5s3rwz7djpr5a25f7xw4mceq"
-	consAddr1, _ := sdk.ConsAddressFromBech32(providerConsAddr1)
-	providerConsAddr2 := "cosmosvalcons1nx7n5uh0ztxsynn4sje6eyq2ud6rc6klc96w39"
-	consAddr2, _ := sdk.ConsAddressFromBech32(providerConsAddr2)
-
-	providerKeeper.UpdateDenylist(ctx, consumerId, []string{providerConsAddr1, providerConsAddr2})
-
-	expectedDenylist := []providertypes.ProviderConsAddress{
-		providertypes.NewProviderConsAddress(consAddr1),
-		providertypes.NewProviderConsAddress(consAddr2)}
-	require.Equal(t, expectedDenylist, providerKeeper.GetDenyList(ctx, consumerId))
-}
-
-func TestUpdateMinimumPowerInTopN(t *testing.T) {
-	providerKeeper, ctx, ctrl, mocks := testkeeper.GetProviderKeeperAndCtx(t, testkeeper.NewInMemKeeperParams(t))
-	defer ctrl.Finish()
-
-	consumerId := "0"
-
-	// test case where Top N is 0 in which case there's no minimum power in top N
-	providerKeeper.SetConsumerPowerShapingParameters(ctx, consumerId, providertypes.PowerShapingParameters{
-		Top_N: 0,
-	})
-
-	err := providerKeeper.UpdateMinimumPowerInTopN(ctx, consumerId, 0, 0)
-	require.NoError(t, err)
-	_, found := providerKeeper.GetMinimumPowerInTopN(ctx, consumerId)
-	require.False(t, found)
-
-	// test cases where Top N > 0 and for this we mock some validators
-	powers := []int64{10, 20, 30}
-	validators := []stakingtypes.Validator{
-		createStakingValidator(ctx, mocks, powers[0], 1), // this validator has ~16 of the total voting power
-		createStakingValidator(ctx, mocks, powers[1], 2), // this validator has ~33% of the total voting gpower
-		createStakingValidator(ctx, mocks, powers[2], 3), // this validator has 50% of the total voting power
-	}
-	mocks.MockStakingKeeper.EXPECT().GetBondedValidatorsByPower(gomock.Any()).Return(validators, nil).AnyTimes()
-
-	maxProviderConsensusValidators := int64(3)
-	params := providerKeeper.GetParams(ctx)
-	params.MaxProviderConsensusValidators = maxProviderConsensusValidators
-	providerKeeper.SetParams(ctx, params)
-
-	// when top N is 50, the minimum power is 30 (because top validator has to validate)
-	providerKeeper.SetConsumerPowerShapingParameters(ctx, consumerId, providertypes.PowerShapingParameters{
-		Top_N: 50,
-	})
-	err = providerKeeper.UpdateMinimumPowerInTopN(ctx, consumerId, 0, 50)
-	require.NoError(t, err)
-	minimumPowerInTopN, found := providerKeeper.GetMinimumPowerInTopN(ctx, consumerId)
-	require.True(t, found)
-	require.Equal(t, int64(30), minimumPowerInTopN)
-
-	// when top N is 51, the minimum power is 20 (because top 2 validators have to validate)
-	providerKeeper.SetConsumerPowerShapingParameters(ctx, consumerId, providertypes.PowerShapingParameters{
-		Top_N: 51,
-	})
-	err = providerKeeper.UpdateMinimumPowerInTopN(ctx, consumerId, 50, 51)
-	require.NoError(t, err)
-	minimumPowerInTopN, found = providerKeeper.GetMinimumPowerInTopN(ctx, consumerId)
-	require.True(t, found)
-	require.Equal(t, int64(20), minimumPowerInTopN)
-
-	// when top N is 100, the minimum power is 10 (that of the validator with the lowest power)
-	providerKeeper.SetConsumerPowerShapingParameters(ctx, consumerId, providertypes.PowerShapingParameters{
-		Top_N: 100,
-	})
-	err = providerKeeper.UpdateMinimumPowerInTopN(ctx, consumerId, 51, 100)
-	require.NoError(t, err)
-	minimumPowerInTopN, found = providerKeeper.GetMinimumPowerInTopN(ctx, consumerId)
-	require.True(t, found)
-	require.Equal(t, int64(10), minimumPowerInTopN)
-}
-
 func TestIsValidatorOptedInToChain(t *testing.T) {
 	providerKeeper, ctx, ctrl, _ := testkeeper.GetProviderKeeperAndCtx(t, testkeeper.NewInMemKeeperParams(t))
 	defer ctrl.Finish()
@@ -643,128 +309,4 @@ func TestIsValidatorOptedInToChain(t *testing.T) {
 	actualConsumerId, found := providerKeeper.IsValidatorOptedInToChainId(ctx, providerAddr, chainId)
 	require.True(t, found)
 	require.Equal(t, expectedConsumerId, actualConsumerId)
-}
-
-func TestPrepareConsumerForLaunch(t *testing.T) {
-	providerKeeper, ctx, ctrl, _ := testkeeper.GetProviderKeeperAndCtx(t, testkeeper.NewInMemKeeperParams(t))
-	defer ctrl.Finish()
-
-	spawnTime := time.Now().UTC()
-	err := providerKeeper.PrepareConsumerForLaunch(ctx, "consumerId", time.Time{}, spawnTime)
-	require.NoError(t, err)
-
-	consumers, err := providerKeeper.GetConsumersToBeLaunched(ctx, spawnTime)
-	require.NoError(t, err)
-	require.Equal(t, providertypes.ConsumerIds{Ids: []string{"consumerId"}}, consumers)
-
-	nextSpawnTime := spawnTime.Add(time.Hour)
-	err = providerKeeper.PrepareConsumerForLaunch(ctx, "consumerId", spawnTime, nextSpawnTime)
-	require.NoError(t, err)
-
-	consumers, err = providerKeeper.GetConsumersToBeLaunched(ctx, spawnTime)
-	require.NoError(t, err)
-	require.Empty(t, consumers)
-
-	consumers, err = providerKeeper.GetConsumersToBeLaunched(ctx, nextSpawnTime)
-	require.NoError(t, err)
-	require.Equal(t, providertypes.ConsumerIds{Ids: []string{"consumerId"}}, consumers)
-}
-
-func TestCanLaunch(t *testing.T) {
-	providerKeeper, ctx, ctrl, _ := testkeeper.GetProviderKeeperAndCtx(t, testkeeper.NewInMemKeeperParams(t))
-	defer ctrl.Finish()
-
-	// cannot launch an unknown chain
-	_, canLaunch := providerKeeper.CanLaunch(ctx, "consumerId")
-	require.False(t, canLaunch)
-
-	// cannot launch a chain without initialization parameters
-	providerKeeper.SetConsumerPhase(ctx, "consumerId", providertypes.ConsumerPhase_CONSUMER_PHASE_INITIALIZED)
-	_, canLaunch = providerKeeper.CanLaunch(ctx, "consumerId")
-	require.False(t, canLaunch)
-
-	// set valid initialization parameters
-	initializationParameters := testkeeper.GetTestInitializationParameters()
-	err := providerKeeper.SetConsumerInitializationParameters(ctx, "consumerId", initializationParameters)
-	require.NoError(t, err)
-
-	// cannot launch a launched chain
-	providerKeeper.SetConsumerPhase(ctx, "consumerId", providertypes.ConsumerPhase_CONSUMER_PHASE_LAUNCHED)
-	_, canLaunch = providerKeeper.CanLaunch(ctx, "consumerId")
-	require.False(t, canLaunch)
-
-	// cannot launch a stopped chain
-	providerKeeper.SetConsumerPhase(ctx, "consumerId", providertypes.ConsumerPhase_CONSUMER_PHASE_STOPPED)
-	_, canLaunch = providerKeeper.CanLaunch(ctx, "consumerId")
-	require.False(t, canLaunch)
-
-	// initialized chain can launch
-	providerKeeper.SetConsumerPhase(ctx, "consumerId", providertypes.ConsumerPhase_CONSUMER_PHASE_INITIALIZED)
-	_, canLaunch = providerKeeper.CanLaunch(ctx, "consumerId")
-	require.True(t, canLaunch)
-
-	// chain cannot launch without a genesis hash
-	initializationParameters.GenesisHash = nil
-	err = providerKeeper.SetConsumerInitializationParameters(ctx, "consumerId", initializationParameters)
-	_, canLaunch = providerKeeper.CanLaunch(ctx, "consumerId")
-	require.NoError(t, err)
-	require.False(t, canLaunch)
-}
-
-func TestHasAtMostOnceCorrectMsgUpdateConsumer(t *testing.T) {
-	providerKeeper, ctx, ctrl, _ := testkeeper.GetProviderKeeperAndCtx(t, testkeeper.NewInMemKeeperParams(t))
-	defer ctrl.Finish()
-
-	expectedMsgUpdateConsumer := providertypes.MsgUpdateConsumer{Signer: "signer", ConsumerId: "consumerId", NewOwnerAddress: "new owner address"}
-
-	proposal, err := govv1.NewProposal([]sdk.Msg{&expectedMsgUpdateConsumer}, 1, time.Now(), time.Now().Add(1*time.Hour), "metadata", "title", "summary", sdk.AccAddress{}, false)
-	require.NoError(t, err)
-
-	_, err = providerKeeper.HasAtMostOnceCorrectMsgUpdateConsumer(ctx, &proposal)
-	require.ErrorContains(t, err, "cannot find owner address")
-
-	// set owner address that is not the gov module
-	providerKeeper.SetConsumerOwnerAddress(ctx, "consumerId", "owner address")
-	_, err = providerKeeper.HasAtMostOnceCorrectMsgUpdateConsumer(ctx, &proposal)
-	require.ErrorContains(t, err, "is not the gov module")
-
-	// set owner address that is the gov module
-	providerKeeper.SetConsumerOwnerAddress(ctx, "consumerId", "cosmos10d07y265gmmuvt4z0w9aw880jnsr700j6zn9kn")
-	actualMsgUpdateConsumer, err := providerKeeper.HasAtMostOnceCorrectMsgUpdateConsumer(ctx, &proposal)
-	require.NoError(t, err)
-	require.Equal(t, expectedMsgUpdateConsumer, *actualMsgUpdateConsumer)
-
-	// a proposal with 2 `MsgUpdateConsumer` messages
-	invalidProposal, err := govv1.NewProposal([]sdk.Msg{&expectedMsgUpdateConsumer, &expectedMsgUpdateConsumer}, 1, time.Now(), time.Now().Add(1*time.Hour), "metadata", "title", "summary", sdk.AccAddress{}, false)
-	require.NoError(t, err)
-	actualMsgUpdateConsumer, err = providerKeeper.HasAtMostOnceCorrectMsgUpdateConsumer(ctx, &invalidProposal)
-	require.ErrorContains(t, err, "proposal can contain at most one")
-	require.Nil(t, actualMsgUpdateConsumer)
-}
-
-func TestDoesNotHaveDeprecatedMessage(t *testing.T) {
-	msgConsumerAddition := providertypes.MsgConsumerAddition{}
-	proposal, err := govv1.NewProposal([]sdk.Msg{&msgConsumerAddition}, 1, time.Now(), time.Now().Add(1*time.Hour), "metadata", "title", "summary", sdk.AccAddress{}, false)
-	require.NoError(t, err)
-	err = keeper.DoesNotHaveDeprecatedMessage(&proposal)
-	require.ErrorContains(t, err, "cannot contain deprecated `MsgConsumerAddition`")
-
-	msgConsumerModification := providertypes.MsgConsumerModification{}
-	proposal, err = govv1.NewProposal([]sdk.Msg{&msgConsumerModification}, 1, time.Now(), time.Now().Add(1*time.Hour), "metadata", "title", "summary", sdk.AccAddress{}, false)
-	require.NoError(t, err)
-	err = keeper.DoesNotHaveDeprecatedMessage(&proposal)
-	require.ErrorContains(t, err, "cannot contain deprecated `MsgConsumerModification`")
-
-	msgConsumerRemoval := providertypes.MsgConsumerRemoval{}
-	proposal, err = govv1.NewProposal([]sdk.Msg{&msgConsumerRemoval}, 1, time.Now(), time.Now().Add(1*time.Hour), "metadata", "title", "summary", sdk.AccAddress{}, false)
-	require.NoError(t, err)
-	err = keeper.DoesNotHaveDeprecatedMessage(&proposal)
-	require.ErrorContains(t, err, "cannot contain deprecated `MsgConsumerRemoval`")
-
-	// a proposal with no deprecated messages
-	msgUpdateConsumer := providertypes.MsgUpdateConsumer{Signer: "signer", ConsumerId: "consumerId", NewOwnerAddress: "new owner address"}
-	proposal, err = govv1.NewProposal([]sdk.Msg{&msgUpdateConsumer}, 1, time.Now(), time.Now().Add(1*time.Hour), "metadata", "title", "summary", sdk.AccAddress{}, false)
-	require.NoError(t, err)
-	err = keeper.DoesNotHaveDeprecatedMessage(&proposal)
-	require.Nil(t, err)
 }
