+++ x/ccv/provider/types/events.go
@@ -4,8 +4,7 @@ package types
 const (
 	EventTypeConsumerClientCreated     = "consumer_client_created"
 	EventTypeAssignConsumerKey         = "assign_consumer_key"
-	EventTypeAddConsumerRewardDenom    = "add_consumer_reward_denom"
-	EventTypeRemoveConsumerRewardDenom = "remove_consumer_reward_denom"
+	EventTypeChangeConsumerRewardDenom = "change_consumer_reward_denom"
 	EventTypeExecuteConsumerChainSlash = "execute_consumer_chain_slash"
 	EventTypeSetConsumerCommissionRate = "set_consumer_commission_rate"
 	EventTypeOptIn                     = "opt_in"
@@ -16,7 +15,8 @@ const (
 	AttributeUnbondingPeriod           = "unbonding_period"
 	AttributeProviderValidatorAddress  = "provider_validator_address"
 	AttributeConsumerConsensusPubKey   = "consumer_consensus_pub_key"
-	AttributeConsumerRewardDenom       = "consumer_reward_denom"
+	AttributeAddConsumerRewardDenom    = "add_consumer_reward_denom"
+	AttributeRemoveConsumerRewardDenom = "remove_consumer_reward_denom"
 	AttributeConsumerCommissionRate    = "consumer_commission_rate"
 	AttributeConsumerId                = "consumer_chain_id"
 )
