+++ x/ccv/provider/types/legacy_proposal_test.go
@@ -542,76 +542,6 @@ func TestChangeRewardDenomsProposalValidateBasic(t *testing.T) {
 	}
 }
 
-func TestMsgUpdateConsumerValidateBasic(t *testing.T) {
-	testCases := []struct {
-		name                   string
-		powerShapingParameters types.PowerShapingParameters
-		expPass                bool
-	}{
-		{
-			"success",
-			types.PowerShapingParameters{
-				Top_N:              50,
-				ValidatorsPowerCap: 100,
-				ValidatorSetCap:    34,
-				Allowlist:          []string{"cosmosvalcons1arwg2sh5fd2wq4mwl6d05pyz4nrwls2mayg36j"},
-				Denylist:           nil,
-				MinStake:           0,
-				AllowInactiveVals:  false,
-			},
-			true,
-		},
-		{
-			"top N is invalid",
-			types.PowerShapingParameters{
-				Top_N:              10,
-				ValidatorsPowerCap: 0,
-				ValidatorSetCap:    0,
-				Allowlist:          nil,
-				Denylist:           nil,
-			},
-			false,
-		},
-		{
-			"validators power cap is invalid",
-			types.PowerShapingParameters{
-				Top_N:              50,
-				ValidatorsPowerCap: 101,
-				ValidatorSetCap:    0,
-				Allowlist:          nil,
-				Denylist:           nil,
-				MinStake:           0,
-				AllowInactiveVals:  false,
-			},
-			false,
-		},
-		{
-			"valid proposal",
-			types.PowerShapingParameters{
-				Top_N:              54,
-				ValidatorsPowerCap: 92,
-				ValidatorSetCap:    0,
-				Allowlist:          []string{"cosmosvalcons1arwg2sh5fd2wq4mwl6d05pyz4nrwls2mayg36j"},
-				Denylist:           []string{"cosmosvalcons1arwg2sh5fd2wq4mwl6d05pyz4nrwls2mayg36j", "cosmosvalcons1arwg2sh5fd2wq4mwl6d05pyz4nrwls2mayg36j"},
-				MinStake:           0,
-				AllowInactiveVals:  false,
-			},
-			true,
-		},
-	}
-
-	for _, tc := range testCases {
-		// TODO (PERMISSIONLESS) add more tests
-		msg, _ := types.NewMsgUpdateConsumer("", "0", "cosmos1p3ucd3ptpw902fluyjzhq3ffgq4ntddac9sa3s", nil, nil, &tc.powerShapingParameters)
-		err := msg.ValidateBasic()
-		if tc.expPass {
-			require.NoError(t, err, "valid case: %s should not return error. got %w", tc.name, err)
-		} else {
-			require.Error(t, err, "invalid case: '%s' must return error but got none", tc.name)
-		}
-	}
-}
-
 func TestValidatePSSFeatures(t *testing.T) {
 	require.NoError(t, types.ValidatePSSFeatures(0, 0))
 	require.NoError(t, types.ValidatePSSFeatures(50, 0))
